Update deactivate button and move search and tabbedsearch files
