+++ mcweb/frontend/src/App.jsx
@@ -23,7 +23,6 @@ import CreateCollection from './features/collections/CreateCollection';
 import CreateSource from './features/sources/CreateSource';
 import CollectionHeader from './features/collections/CollectionHeader';
 import GeographicNewsCollections from './features/collections/GeographicNewsCollections';
-import Search from './features/search/Search';
 import SourceShow from './features/sources/SourceShow';
 import ListSourceFeeds from './features/sources/ListSourceFeeds';
 import CreateFeed from './features/feeds/CreateFeed';
@@ -32,7 +31,7 @@ import FeedHeader from './features/feeds/FeedHeader';
 import FeedShow from './features/feeds/FeedShow';
 import SourceHeader from './features/sources/SourceHeader';
 import ReleaseNotes from './features/about/ReleaseNotes';
-import TabbedSearch from './features/search/query/TabbedSearch';
+import TabbedSearch from './features/search/TabbedSearch';
 
 import StoryShow from './features/stories/StoryShow';
 import ModifyCollection from './features/collections/ModifyCollection';
