+++ mcweb/frontend/src/features/search/util/deactivateButton.js
@@ -1,53 +1,50 @@
-import dayjs from "dayjs";
+import dayjs from 'dayjs';
 
+const deactvateButton = (queryState) => {
+  let returnVal = true;
+  queryState.forEach((queryObject) => {
+    const {
+      queryString,
+      queryList,
+      negatedQueryList,
+      startDate,
+      endDate,
+    } = queryObject;
 
-const deactvateButton = (queryObject) => {
-  const {
-    queryString,
-    queryList,
-    negatedQueryList,
-    startDate,
-    endDate,
-  } = queryObject;
-  
-  
-  
-  const totalQuery = queryList.concat(negatedQueryList)
-  
-  // is the query string empty? 
-  const isQueryEmpty = validQuery(totalQuery)
+    const totalQuery = queryList.concat(negatedQueryList);
 
-  // are the dates in correct order? 
-  const areDatesValid = validDates(startDate, endDate); 
-
-  // is the advanced search query empty? 
-  const isQueryStringEmpty = validQueryString(queryString)
-  
-
-
-  return ((isQueryEmpty || isQueryStringEmpty) && areDatesValid)
-  
-}
-
-// checks too see if the query is empty 
-function validQuery(totalQuery) {
-  for(let i = 0; i < totalQuery.length; i++) {
-    if(totalQuery[i].length > 0) {
-      return true; 
+    // checks too see if the query is empty
+    function validQuery(tQ) {
+      for (let i = 0; i < tQ.length; i += 1) {
+        if (tQ[i].length > 0) {
+          return true;
+        }
+      }
+      return false;
     }
-  }
-  return false; 
-}
 
-// checks to see if the startDAte is before the endDAte
-function validDates(startDate, endDate) {
-  return dayjs(startDate).isBefore(dayjs(endDate));
-}
+    // checks to see if the startDAte is before the endDAte
+    function validDates(sD, eD) {
+      return dayjs(sD).isBefore(dayjs(eD));
+    }
 
-// is the advanced search query string not just the "*"
-function validQueryString(queryString) {
-  return queryString.length != 0; 
-}
+    // is the advanced search query string not just the "*"
+    function validQueryString(qS) {
+      return qS.length !== 0;
+    }
 
+    // is the query string empty?
+    const isQueryEmpty = validQuery(totalQuery);
 
-export default deactvateButton
+    // are the dates in correct order?
+    const areDatesValid = validDates(startDate, endDate);
+
+    // is the advanced search query empty?
+    const isQueryStringEmpty = validQueryString(queryString);
+
+    returnVal = ((isQueryEmpty || isQueryStringEmpty) && areDatesValid);
+  });
+  return returnVal;
+};
+
+export default deactvateButton;
